Bug 584141 & 584157 changes

+++ ElitaPlus/Claims/ClaimWizardForm.aspx.vb
@@ -442,11 +442,14 @@ Public Class ClaimWizardForm
                         ' REQ- 6156 - Skipping this step of choosing Service Center and copied the required code to step 5
 
                         For Each claimAuth As ClaimAuthorization In Me.State.ClaimBO.ClaimAuthorizationChildren
-                            If Not claimAuth.IsNew Then
-                                claimAuth.Void()
-                            Else
-                                claimAuth.DeleteChildren()
-                                claimAuth.Delete()
+                            'below line is to skip voiding or deleting authorizations if they are of claim deductible refund related
+                            If Not claimAuth.IsAuthorizationDeductibleRefund Then
+                                If Not claimAuth.IsNew Then
+                                    claimAuth.Void()
+                                Else
+                                    claimAuth.DeleteChildren()
+                                    claimAuth.Delete()
+                                End If
                             End If
                         Next
                         ' REQ- 6156 - Call Start Fulfillment process web method in Fulfillment Web Service
@@ -463,11 +466,14 @@ Public Class ClaimWizardForm
                         If Not attvalue Is Nothing AndAlso attvalue.Value = Codes.YESNO_Y Then
                             If (Me.State.ClaimBO.Status = BasicClaimStatus.Active) Then
                                 For Each claimAuth As ClaimAuthorization In Me.State.ClaimBO.ClaimAuthorizationChildren
-                                    If Not claimAuth.IsNew Then
-                                        claimAuth.Void()
-                                    Else
-                                        claimAuth.DeleteChildren()
-                                        claimAuth.Delete()
+                                    'below line is to skip voiding or deleting authorizations if they are of claim deductible refund related
+                                    If Not claimAuth.IsAuthorizationDeductibleRefund Then
+                                        If Not claimAuth.IsNew Then
+                                            claimAuth.Void()
+                                        Else
+                                            claimAuth.DeleteChildren()
+                                            claimAuth.Delete()
+                                        End If
                                     End If
                                 Next
                                 Me.State.DoesActiveTradeInExistForIMEI = DoesAcceptedOfferExistForIMEI()
@@ -1201,11 +1207,6 @@ Public Class ClaimWizardForm
             Case ClaimWizardSteps.Step5
         End Select
 
-        If (Me.State.ClaimBO.IsDeductibleRefundAllowed AndAlso
-            Not Me.State.ClaimBO.IsDeductibleRefundExist) Then
-            ControlMgr.SetVisibleControl(Me, btnClaimDeductibleRefund, True)
-        End If
-
         HandleButtons(wizardStep)
 
     End Sub
@@ -1575,6 +1576,7 @@ Public Class ClaimWizardForm
         ControlMgr.SetVisibleControl(Me, btnSave_WRITE, False)
         ControlMgr.SetVisibleControl(Me, btnClaimOverride_Write, False)
         ControlMgr.SetVisibleControl(Me, btnComment, False)
+        ControlMgr.SetVisibleControl(Me, btnClaimDeductibleRefund, False)
         Me.btnContinue.Text = TranslationBase.TranslateLabelOrMessage("CONTINUE")
 
         Select Case wizardStep
@@ -1606,6 +1608,10 @@ Public Class ClaimWizardForm
                 ControlMgr.SetVisibleControl(Me, btnDenyClaim, Not Me.State.IsEditMode)
                 ControlMgr.SetVisibleControl(Me, Me.btnContinue, Not Me.State.InputParameters.ComingFromDenyClaim)
 
+                If (Me.State.ClaimBO.IsDeductibleRefundAllowed AndAlso
+                            Not Me.State.ClaimBO.IsDeductibleRefundExist) Then
+                    ControlMgr.SetVisibleControl(Me, btnClaimDeductibleRefund, True)
+                End If
             Case ClaimWizardSteps.Step4
             Case ClaimWizardSteps.Step5
                 Me.btnContinue.Text = TranslationBase.TranslateLabelOrMessage("SUBMIT_CLAIM")
@@ -2509,17 +2515,17 @@ Public Class ClaimWizardForm
                     .ClaimBO = CType(State.ClaimBO, ClaimBase)
                     If Not allowEnrolledDeviceUpdate Is Nothing AndAlso allowEnrolledDeviceUpdate.Value = Codes.YESNO_Y Then
                         For Each i As ClaimIssue In State.ClaimBO.ClaimIssuesList
-                            If i.IssueCode = ISSUE_CODE_CR_DEVICE_MIS and i.StatusCode = Codes.CLAIMISSUE_STATUS__OPEN Then
+                            If i.IssueCode = ISSUE_CODE_CR_DEVICE_MIS And i.StatusCode = Codes.CLAIMISSUE_STATUS__OPEN Then
                                 .ShowDeviceEditImg = True
                                 Exit For
-                            Else 
+                            Else
                                 .ShowDeviceEditImg = False
                             End If
                         Next
-                    else
-                       .ShowDeviceEditImg = False
+                    Else
+                        .ShowDeviceEditImg = False
                     End If
-                    
+
                 End With
             End If
         End With
@@ -3970,14 +3976,7 @@ Public Class ClaimWizardForm
 
     Private Sub btnClaimDeductibleRefund_Click(sender As Object, e As EventArgs) Handles btnClaimDeductibleRefund.Click
         Try
-            Dim FlowName As String = CLAIM_WIZARD_FROM_EVENT_DETAILS
-            Dim nav As New ElitaPlusNavigation
-            Me.NavController = New NavControllerBase(nav.Flow(FlowName))
-            Me.NavController.State = New MyState
-            Me.NavController.Navigate(Me, FlowEvents.EVENT_CLAIM_DEDUCTIBLE_REFUND, New ClaimDeductibleRefundForm.Parameters(CType(Me.State.ClaimBO, ClaimBase)))
-
-        Catch ex As Threading.ThreadAbortException
-
+            Me.callPage(ClaimDeductibleRefundForm.URL, New ClaimDeductibleRefundForm.Parameters(CType(Me.State.ClaimBO, ClaimBase)))
         Catch ex As Exception
             Me.HandleErrors(ex, Me.MasterPage.MessageController)
         End Try
